getMoleFractions returns all zeros when total amount drops too low.
Previously, when the amount fell below machine precision, the largest would 
still have a mole fraction of 1, and thus still have a high vapor pressure,
even though there was none of it left. This is what led to the continued
evaporation of the least volatile component, even when it was all gone,
leading to the negative amounts.